Starting things off
Updated the readme with some examples of the intended use-case.

Added most of the important classes that will be used and critical functions that will need to be implemented first.

MathLib will handle most of the heavy lifting.
Included an example case of implicit operator overloading for addition on the Term class.