readme update
Added C# formatting to the codeblocks.
Added a short message under the description of the readme. Since this is a public repo and the scope for this project is very large, not all sections may be completed yet and may take time to fully build for the most basic uses.